zuowen/requirements.txt
flask
dirtyjson
celery
redis
openai
python-dotenv
fastapi
uvicorn[standard]
python-multipart

